refactor: self-invocation 문제 해결을 위해 syncForStadium을 별도 빈으로 분리
